feat(apps/server): move TagsModule out of ProductsModule to a standalone module
